moved crypto stuff to a new utils file and removed xrpBase58.js since we now use base-x to do decoding
