--- utils/xrpBase58.js
@@ -1,69 +1,0 @@
-/**
- * alphabet used:
- * rpshnaf39wBUDNEGHJKLM4PQRST7VWXYZ2bcdeCg65jkm8oFqi1tuvAxyz
- * 
- * https://en.wikipedia.org/wiki/Base58
- */
-const symbolMap = {
-		r: 0,
-		p: 1,
-		s: 2,
-		h: 3,
-		n: 4,
-		a: 5,
-		f: 6,
-		3: 7,
-		9: 8,
-		w: 9,
-		B: 10,
-		U: 11,
-		D: 12,
-		N: 13,
-		E: 14,
-		G: 15,
-		H: 16, 
-		J: 17,
-		K: 18,
-		L: 19,
-		M: 20,
-		4: 21, 
-		P: 22,
-		Q: 23,
-		R: 24,
-		S: 25, 
-		T: 26,
-		7: 27,
-		V: 28,
-		W: 29,
-		X: 30,
-		Y: 31,
-		Z: 32,
-		2: 33,
-		b: 34,
-		c: 35,
-		d: 36,
-		e: 37,
-		C: 38,
-		g: 39,
-		6: 40,
-		5: 41,
-		j: 42,
-		k: 43,
-		m: 44,
-		8: 45,
-		o: 46,
-		F: 47,
-		q: 48,
-		i: 49,
-		1: 50,
-		t: 51,
-		u: 52,
-		v: 53,
-		A: 54,
-		x: 55,
-		y: 56,
-		z: 57
-}
-module.exports = {
-		
-}
